requirements.txt
pygame==2.5.2
numpy>=1.26
PyOpenGL==3.1.7
